docs: Display the correct information for the Google Markers for each Place listed.
app.js: Contains the functionality of the Google map and listed places. The Google map loads using the Google API and some of its 'Drawing on the Map' tools. The listed places are observed by Knockoutjs observable functionality. Foursquare API is used to load basic information of the each place listed. The basic info are: Category and Address.

locations.js: Contains an array of the places with title, location (lat and lng), and img path. This array is used for the markers and the list.

Functionality Completed: Knockoutjs loads the list of places, Google API loads the map with 5 markers using the location(lat,lng). Markers displayed title, img, category and address. Basic style was applied.