feat: Styled and added functionality to the Marker's InfoWindow
Each of the the marker's infowindow displays Name of the place, a link to carousel with images, category of food, and physical address.

The name of the place its pull from an array stored inside locations.js file.

The images are pulled from Foursquare API. They are display in a modal window using bootstrap. Only up to 10 images are display per place.

The category, physical address, and images of each place are pull form Foursquare API.

The icons used are from Fontawesome.com. The map markers fro iconfinder.com

Used for infowindow:

- Bootstrap
- Foursquare API
- Google markers
- Fontawesome
- Iconfinder
- CSS Grid