feat: Add Animation to markers and sidenav. Add a input filter and screen responsiveness.
Utilized knockoutjs compute observables to filter the listed places and markers by typing in the input filter. Simple highlight and bounce animation was added to markers for when the markers or the listed places are mouse over.

A sidenav was added with a Javascript function and adapted its width to full screen for mobile and for over 500px to 350px;